reorganization - fancy welcome page - button modif
