+++ src/components/HomePage.vue
@@ -1,12 +1,19 @@
 <template>
-  <div id = "homepage">
-      <router-view> </router-view>
+  <div>
+    <mascotte>
+      <h1> WELCOME !</h1>
+      <img class="mascotte" src="./img/mascotte_00.png" alt="mascotte rhino">
+
+    </mascotte>
   </div>
 </template>
 
 <script>
 export default {
-
+  name: "HomePage"
 }
 </script>
 
+<style scoped>
+
+</style>
