+++ src/components/Map.vue
@@ -1,9 +1,6 @@
 <template>
-  <div>
-    <div class="upper_container">
-      <img class="upper_container" src ="./img/Lugdunum_logo.png" alt="lugdunum logo">
-      <h1 class="upper_container" >Website for Subscription and Statistics</h1>
-    </div>
+  <div id="map">
+    <h1> LUGDUNUM MAP </h1>
   </div>
 
 </template>
