+++ src/index.js
@@ -1,7 +1,7 @@
 import Vue from "vue";
-import App from './components/App.vue';
+import App from './components/LogIn.vue';
 import Map from './components/Map.vue';
-import HomePage from './components/HomePage.vue';
+import HomePage from './components/App.vue';
 
 import ApolloClient from 'apollo-boost';
 import VueApollo from "vue-apollo";
@@ -23,12 +23,16 @@ const router = new VueRouter({
   mode: 'history',
   routes: [
     {
-      path: '/app',
-      component: require('./components/App.vue').default
+      path: '/login',
+      component: require('./components/LogIn.vue').default
     },
     {
       path: '/map',
       component: require('./components/Map.vue').default
+    },
+    {
+      path: '/',
+      component: require('./components/HomePage.vue').default
     },
     {
       //default route
@@ -41,7 +45,7 @@ const router = new VueRouter({
 new Vue({
   apolloProvider,
   router: router,
-  render: h => h(require('./components/HomePage.vue').default)
+  render: h => h(require('./components/App.vue').default)
 }).$mount('#app')
