+++ src/stylesheet/style.css
@@ -1,12 +1,13 @@
 body{
-  background-color: chocolate;
+  background-image: url("../components/img/background_00.PNG");
+  background-size: 100%;
   flex-direction: column;
-  justify-content: left;
+  justify-content: center;
 }
-
+/* Upper Container ==================================================================================*/
 div.upper_container {
   padding: 0.5%;
-  background-color: sienna;
+  background-color: #a0522d99;
   width: 100%;
   height: 5%;
   display: flex;
@@ -18,14 +19,56 @@ img.upper_container {
   align-self: center;
   width: 10%;
 }
+img.upper_container:hover {
+  transform: scale(1.05);
+}
 
-h1.upper_container {
+a.upper_container{
   align-self: center;
+  font-family: 'Open-Sans', sans-serif;
+  font-size:1.2vw;
+  color: black;
+  padding: 10px;
+}
+nav{
+  float: right;
+  align-self: center;
+  padding-left: 20%;
+}
+
+/* button======================================================== */
+input[type=button]{
   font-family: 'Open-Sans', sans-serif;
   font-size:1.5vw;
 }
+.button {
+  display: inline-block;
+  align-self: center;
+  border: none;
+  border-radius: 15px;
+  outline: none;
+  background-color: #d2691e;
+  color: black;
+  font-size: 1.3vw;
+  text-align: center;
+  padding-left: 15px;
+  padding-right: 15px;
+  transition: all 0.5s;
+  cursor: pointer;
+  margin: 0 5px;
+  box-shadow: 0 5px #000000B0;
+}
 
-nav {
+.button:hover {background-color: #b85d19
+}
+
+.button:active {
+  background-color: #b85d19;
+  box-shadow: 0 2px #000000B0;
+  transform: translateY(8px);
+}
+
+loginStuff {
   width: 50%;
   padding: 0.2%;
   background-color: darkorange;
@@ -34,47 +77,36 @@ nav {
   justify-content: space-between;
 }
 
-input {
+input[type=input] {
+  height: 40%;
+  align-self: center;
+  padding-left: 15px;
+  padding-right: 15px;
+  font-family: 'Open-Sans', sans-serif;
+  font-size: 1vw;
   background-color: transparent;
 }
 
-/* button======================================================== */
-button {
-  height: 80%;
-  background-color: antiquewhite;
-  border: 2px solid darkblue;
-  color: darkblue;
-  text-align: center;
+h1 {
+  align-self: center;
   font-family: 'Open-Sans', sans-serif;
   font-size:1.5vw;
-  padding-left: 15px;
-  padding-right: 15px;
-  transition: all 0.5s;
-  cursor: pointer;
-  margin: 0 5px;
+  padding-right: 25%;
 }
 
-button span {
-  cursor: pointer;
-  display: inline-block;
-  position: relative;
-  transition: 0.5s;
-}
-
-button span:after {
-  content: '\00bb';
-  position: absolute;
-  opacity: 0;
-  top: 0;
-  right: -20px;
-  transition: 0.5s;
-}
-
-button:hover span {
-  padding-right: 25px;
-}
-
-button:hover span:after {
-  opacity: 1;
+mascotte {
+  position: fixed;
+  bottom: 0;
   right: 0;
+  width: 20%;
+  height: 70%;
 }
+img.mascotte{
+   display: flex;
+   justify-content: space-evenly;
+   align-self: center;
+   height: 800px;
+}
+
+
+#loadPage{display: none;}
